backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
passlib[bcrypt]
bcrypt==4.0.1
python-jose
python-multipart
python-dotenv
pydantic
google-generativeai